# Conflicts:
#	src/Controller/MainController.php
#	webpack.config.js

+++ src/Controller/SortieController.php
@@ -3,8 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\Campus;
+use App\Entity\Etat;
+use DateTime;
 use App\Entity\Lieu;
 use App\Entity\Sortie;
+use App\Form\AnnulerSortieType;
+use App\Form\CreationSortieType;
 use App\Form\SortieType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -20,6 +24,16 @@ class SortieController extends AbstractController
         $sortie=$entityManager->getRepository(Sortie::class)->find($id);
         if(!$sortie) {
             $this->addFlash('fail', 'Sortie n\'existe pas !');
+            return $this->redirectToRoute('main_home');
+        }
+        if ($this->getUser() !== $sortie->getOrganisateur()){
+            $this->addFlash('fail', ' Vous n\'êtes pas l\'organisateur de cette Sortie !');
+            return $this->redirectToRoute('main_home');
+        }
+        $etat= new Etat();
+        $etat=$entityManager->getRepository($etat::class)->findOneBy(['libelle'=>'Créée']);
+        if ($sortie->getEtat() !== $etat  ){
+            $this->addFlash('fail', ' Cette sortie n\'est plus modifiable !');
             return $this->redirectToRoute('main_home');
         }
         $sortieForm = $this->createForm(SortieType::class, $sortie);
@@ -40,14 +54,30 @@ class SortieController extends AbstractController
 
                 $entityManager->persist($sortie);
                 $entityManager->flush();
-                $this->addFlash('success', 'Sortie modifiée avec succès !');
+                $this->addFlash('success', 'Sortie a été modifiée avec succès !');
                 return $this->redirectToRoute('main_home');
-            }else {
+            }elseif ($sortieForm->get('supprimer')->isClicked()) {
                 $sortie = $sortieForm->getData();
                 $entityManager->remove($sortie);
                 $entityManager->flush();
-                $this->addFlash('success', 'Sortie suprimée avec succès !');
+                $this->addFlash('success', 'Sortie a été supprimée avec succès !');
                 return $this->redirectToRoute('main_home');
+            } else {
+                $sortie = $sortieForm->getData();
+                $latitude = $sortieForm->get('latitude')->getData();
+                $longitude = $sortieForm->get('longitude')->getData();
+                $lieu = $sortie->getLieu();
+                $lieu->setLatitude($latitude);
+                $lieu->setLongitude($longitude);
+                $sortie->setLieu($lieu);
+                 $etat= new Etat();
+                 $etat=$entityManager->getRepository($etat::class)->findOneBy(['libelle'=>'Ouverte']);
+                 $sortie->setEtat($etat);
+                $entityManager->persist($sortie);
+                $entityManager->flush();
+                $this->addFlash('success', 'Sortie a été publiée avec succès !');
+                return $this->redirectToRoute('main_home');
+
             }
         }
 
@@ -57,13 +87,73 @@ class SortieController extends AbstractController
         ]);
     }
 
+    #[Route('/creerSortie', name: 'app_creation_sortie')]
+    public function creation(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $sortie= new Sortie();
+        $creationSortieForm = $this->createForm(CreationSortieType::class, $sortie);
+        $creationSortieForm->handleRequest($request);
 
-    #[Route('/AfficherSortie/{id}', name: 'app_afficherSortie')]
+        if ($creationSortieForm->isSubmitted() && $creationSortieForm->isValid()){
+
+            if ($creationSortieForm->get('enregistrer')->isClicked()) {
+                $sortie = $creationSortieForm->getData();
+                $latitude = $creationSortieForm->get('latitude')->getData();
+                $longitude = $creationSortieForm->get('longitude')->getData();
+                $lieu = $sortie->getLieu();
+                $lieu->setLatitude($latitude);
+                $lieu->setLongitude($longitude);
+                $sortie->setLieu($lieu);
+                $user=$this->getUser();
+                $sortie->setOrganisateur($user);
+                $etat= new Etat();
+                $etat=$entityManager->getRepository($etat::class)->findOneBy(['libelle'=>'Créée']);
+                $sortie->setEtat($etat);
+
+                $entityManager->persist($sortie);
+                $entityManager->flush();
+                $this->addFlash('success', 'Sortie a été crée avec succès !');
+                return $this->redirectToRoute('main_home');
+            } else {
+                $sortie = $creationSortieForm->getData();
+                $latitude = $creationSortieForm->get('latitude')->getData();
+                $longitude = $creationSortieForm->get('longitude')->getData();
+                $lieu = $sortie->getLieu();
+                $lieu->setLatitude($latitude);
+                $lieu->setLongitude($longitude);
+                $sortie->setLieu($lieu);
+                $user=$this->getUser();
+                $sortie->setOrganisateur($user);
+                $etat= new Etat();
+                $etat=$entityManager->getRepository($etat::class)->findOneBy(['libelle'=>'Ouverte']);
+                $sortie->setEtat($etat);
+                $entityManager->persist($sortie);
+                $entityManager->flush();
+                $this->addFlash('success', 'Sortie a été publiée avec succès !');
+                return $this->redirectToRoute('main_home');
+
+            }
+        }
+
+        return $this->render('sortie/creerSortie.html.twig', [
+            'creationSortieForm' => $creationSortieForm->createView(),
+            'sortie' =>$sortie,
+        ]);
+    }
+
+
+    #[Route('/afficherSortie/{id}', name: 'app_afficherSortie')]
     public function afficher(int $id, Request $request, EntityManagerInterface $entityManager): Response
     {
         $sortie=$entityManager->getRepository(Sortie::class)->find($id);
         if(!$sortie) {
             $this->addFlash('fail', 'Sortie n\'existe pas !');
+            return $this->redirectToRoute('main_home');
+        }
+        $unMoisAuparavant = new DateTime();
+        $unMoisAuparavant->modify('-1 month');
+        if($sortie->getDateHeureDebut() <= $unMoisAuparavant) {
+            $this->addFlash('fail', ' Cette sortie n\'est plus accesible !');
             return $this->redirectToRoute('main_home');
         }
         $users=$sortie->getUsers();
@@ -75,6 +165,76 @@ class SortieController extends AbstractController
       /*  dd($users);*/
         return $this->render('sortie/afficherSortie.html.twig', [
                        'sortie' =>$sortie,
+        ]);
+    }
+
+
+    #[Route('/sInscrire/{id}', name: 'app_sInscrire')]
+    public function Sinscrire(int $id, Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $sortie=$entityManager->getRepository(Sortie::class)->find($id);
+        if(!$sortie) {
+            $this->addFlash('fail', 'Sortie n\'existe pas !');
+            return $this->redirectToRoute('main_home');
+        }
+        $user=$this->getUser();
+        $sortie->addUser($user);
+
+        $entityManager->persist($sortie);
+        $entityManager->flush();
+        $this->addFlash('success', 'Vous êtes inscrit avec succès !');
+
+        return $this->redirectToRoute('main_home');
+    }
+
+    #[Route('/seDesister/{id}', name: 'app_seDesister')]
+    public function SeDesister(int $id, Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $sortie=$entityManager->getRepository(Sortie::class)->find($id);
+        if(!$sortie) {
+            $this->addFlash('fail', 'Sortie n\'existe pas !');
+            return $this->redirectToRoute('main_home');
+        }
+        $user=$this->getUser();
+        $sortie->removeUser($user);
+
+        $entityManager->persist($sortie);
+        $entityManager->flush();
+        $this->addFlash('success', 'Vous êtes désinscrit avec succès !');
+
+        return $this->redirectToRoute('main_home');
+    }
+
+
+    #[Route('/annulerSortie/{id}', name: 'app_annuler')]
+    public function annuler(int $id, Request $request, EntityManagerInterface $entityManager): Response
+    {
+        $sortie=$entityManager->getRepository(Sortie::class)->find($id);
+        if(!$sortie) {
+            $this->addFlash('fail', 'Sortie n\'existe pas !');
+            return $this->redirectToRoute('main_home');
+        }
+        if ($this->getUser() !== $sortie->getOrganisateur()){
+            $this->addFlash('fail', ' Vous n\'êtes pas l\'organisateur de cette Sortie !');
+            return $this->redirectToRoute('main_home');
+        }
+        $annulerSortieForm = $this->createForm(AnnulerSortieType::class, $sortie);
+        $annulerSortieForm->handleRequest($request);
+        if ($annulerSortieForm->isSubmitted() && $annulerSortieForm->isValid()){
+
+            $sortie = $annulerSortieForm->getData();
+            $etat= new Etat();
+            $etat=$entityManager->getRepository($etat::class)->findOneBy(['libelle'=>'Annulée']);
+            $sortie->setEtat($etat);
+
+            $entityManager->persist($sortie);
+            $entityManager->flush();
+            $this->addFlash('success', 'Sortie annulée avec succès !');
+            return $this->redirectToRoute('main_home');
+            }
+        return $this->render('sortie/annulerSortie.html.twig', [
+            'annulerSortieForm' => $annulerSortieForm->createView(),
+            'sortie' =>$sortie,
         ]);
     }
 }
